+++ src/app/mentor/profile/page.tsx
@@ -1,0 +1,5 @@
+import { ProfileForm } from '@/components/ProfileForm';
+
+export default function MentorProfilePage() {
+  return <ProfileForm role="MENTOR" />;
+}
