+++ src/i18n/dictionaries.ts
@@ -1712,6 +1712,7 @@ const en = {
     phone: 'Phone Number',
     whatsapp: 'WhatsApp',
     city: 'City',
+    cityHint: 'e.g. Monheim',
     birthDate: 'Birth Date',
     about: 'About you',
     displayName: 'Display name',
@@ -1727,12 +1728,20 @@ const en = {
     interestsHint: 'e.g. AI, web, fintech',
     mentorCapacity: 'Mentee capacity',
     mentorCapacityHint: 'How many mentees you can take on at once',
+    languages: 'Languages spoken',
+    languagesPlaceholder: 'e.g. English, Turkish, German',
+    languagesHint: 'Separate multiple languages with commas',
     education: 'Education',
     university: 'University',
+    universityHint: 'e.g. MIT',
     department: 'Department / Major',
+    departmentHint: 'e.g. Computer Science',
     graduationYear: 'Graduation Year',
+    graduationYearNotSet: 'Not set',
     skillsCv: 'Skills & CV',
     skills: 'Skills',
+    skillsPlaceholder: 'React, Python, Data Analysis',
+    skillsHint: 'Separate multiple skills with commas',
     cvUrl: 'CV URL',
     cvUrlHint: 'Optional: link to an external CV (Drive, etc.). Or upload a file below.',
     profileViews: 'Profile views',
@@ -3535,6 +3544,7 @@ const tr: Dict = {
     phone: 'Telefon',
     whatsapp: 'WhatsApp',
     city: 'Şehir',
+    cityHint: 'örn. Monheim',
     birthDate: 'Doğum Tarihi',
     about: 'Hakkında',
     displayName: 'Görünen ad',
@@ -3550,12 +3560,20 @@ const tr: Dict = {
     interestsHint: 'örn. yapay zeka, web, fintech',
     mentorCapacity: 'Mentee kapasitesi',
     mentorCapacityHint: 'Aynı anda kaç mentee alabilirsin',
+    languages: 'Konuşulan diller',
+    languagesPlaceholder: 'örn. Türkçe, İngilizce, Almanca',
+    languagesHint: 'Birden fazla dili virgülle ayır',
     education: 'Eğitim',
     university: 'Üniversite',
+    universityHint: 'örn. İTÜ',
     department: 'Bölüm',
+    departmentHint: 'örn. Bilgisayar Mühendisliği',
     graduationYear: 'Mezuniyet Yılı',
+    graduationYearNotSet: 'Belirtilmedi',
     skillsCv: 'Yetenekler & CV',
     skills: 'Yetenekler',
+    skillsPlaceholder: 'React, Python, Veri Analizi',
+    skillsHint: 'Birden fazla yeteneği virgülle ayır',
     cvUrl: 'CV Bağlantısı',
     cvUrlHint: 'İsteğe bağlı: harici bir CV bağlantısı (Drive vb.). Ya da aşağıdan dosya yükleyin.',
     profileViews: 'Profil görüntülenme',
@@ -5356,6 +5374,7 @@ const de: Dict = {
     phone: 'Telefonnummer',
     whatsapp: 'WhatsApp',
     city: 'Stadt',
+    cityHint: 'z. B. Monheim',
     birthDate: 'Geburtsdatum',
     about: 'Über dich',
     displayName: 'Anzeigename',
@@ -5371,12 +5390,20 @@ const de: Dict = {
     interestsHint: 'z. B. KI, Web, Fintech',
     mentorCapacity: 'Mentee-Kapazität',
     mentorCapacityHint: 'Wie viele Mentees du gleichzeitig betreuen kannst',
+    languages: 'Gesprochene Sprachen',
+    languagesPlaceholder: 'z. B. Deutsch, Englisch, Türkisch',
+    languagesHint: 'Mehrere Sprachen durch Kommas trennen',
     education: 'Ausbildung',
     university: 'Universität',
+    universityHint: 'z. B. TU Berlin',
     department: 'Fachbereich / Studienfach',
+    departmentHint: 'z. B. Informatik',
     graduationYear: 'Abschlussjahr',
+    graduationYearNotSet: 'Nicht angegeben',
     skillsCv: 'Skills & Lebenslauf',
     skills: 'Skills',
+    skillsPlaceholder: 'React, Python, Datenanalyse',
+    skillsHint: 'Mehrere Skills durch Kommas trennen',
     cvUrl: 'Lebenslauf-URL',
     cvUrlHint: 'Optional: Link zu einem externen Lebenslauf (Drive usw.). Oder unten eine Datei hochladen.',
     profileViews: 'Profilaufrufe',
